Missing "Array.prototype" fails instanceof Array check
Performing queries with an array in the value causes a failure in the Firestore serializer where "val instanceof Array" returns false due to "Array.prototype" no longer being on the object's [[Prototype]] chain once the "msg" is being processed by the contrib. 

Array.isArray can be used to properly detect an Array so it can be recreated prior to calling the Firebase "where" function.